Making some final refactoring in the imports to reflect the correct folders
