+++ MontyHallKata/Views/InputOutput/IInputOutput.cs
@@ -1,4 +1,4 @@
-namespace MontyHallKata.Views.Console
+namespace MontyHallKata.Views.InputOutput
 {
     public interface IInputOutput 
     {
