+++ MontyHallTests/ControllersTest/GameplayTests.cs
@@ -7,7 +7,7 @@ using MontyHallKata.Models.Randomizer;
 using Moq;
 using Xunit;
 
-namespace MontyHallTests
+namespace MontyHallTests.ControllersTest
 {
     public class GameplayTests
     {
